Fix test organisation
